# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go-Rocks is a tech playground with Go, Gin",
        "title": "Swagger Go-Rocks",
        "contact": {
            "name": "Lingga Edo",
            "url": "http://www.github.com/linggaaskaedo",
            "email": "lemp.otis@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/employee": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for get a employee with param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Get list of employee based on query params",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "must-db-revalidate"
                        ],
                        "type": "string",
                        "description": "Request cache control",
                        "name": "Cache-Control",
                        "in": "header"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Search by Employee Public ID",
                        "name": "public_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by Employee first name",
                        "name": "first_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by Employee last name",
                        "name": "last_name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "M",
                            "F"
                        ],
                        "type": "string",
                        "description": "Search by Employee gender",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "-id",
                            "firstname",
                            "-firstname",
                            "lastname",
                            "-lastname",
                            "gender",
                            "-gender"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Sort result by these attributes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPEmployeesResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/employee/{employeeID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for get a employee with ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Get employee by ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "employeeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPEmployeeResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/kc/employee/{employeeID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for get a employee with ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KC"
                ],
                "summary": "Get employee by ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "employeeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPEmployeeResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/kc/login": {
            "post": {
                "description": "Endpoint for generating access token for user via keycloak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KC"
                ],
                "summary": "Keycloak Login",
                "parameters": [
                    {
                        "description": "Payload user login",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPUserLoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.EmployeeDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0"
                },
                "firstname": {
                    "type": "string",
                    "x-order": "1"
                },
                "lastname": {
                    "type": "string",
                    "x-order": "2"
                },
                "gender": {
                    "type": "string",
                    "x-order": "3"
                }
            }
        },
        "dto.Meta": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "x-order": "0"
                },
                "status_code": {
                    "type": "integer",
                    "x-order": "1"
                },
                "status": {
                    "type": "string",
                    "x-order": "2"
                },
                "message": {
                    "type": "string",
                    "x-order": "3"
                },
                "error": {
                    "type": "object",
                    "x-order": "4"
                },
                "timestamp": {
                    "type": "string",
                    "x-order": "5"
                }
            }
        },
...
        "dto.UserLoginDTO": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "x-order": "0"
                },
                "refreshToken": {
                    "type": "string",
                    "x-order": "1"
                },
                "expiresAt": {
                    "type": "integer",
                    "x-order": "2"
                },
                "expiresRt": {
                    "type": "integer",
                    "x-order": "3"
                },
                "expiresIn": {
                    "type": "integer",
                    "x-order": "4"
                }
            }
        },
...
        "entity.Pagination": {
            "type": "object",
            "properties": {
                "current_elements": {
                    "type": "integer"
                },
                "current_page": {
                    "type": "integer"
                },
                "cursor_end": {
                    "type": "string"
                },
                "cursor_start": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "rest.EmployeeData": {
            "type": "object",
            "properties": {
                "employee": {
                    "$ref": "#/definitions/dto.EmployeeDTO"
                }
            }
        },
        "rest.EmployeesData": {
            "type": "object",
            "properties": {
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EmployeeDTO"
                    }
                }
            }
        },
...
        "rest.HTTPEmployeeResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.EmployeeData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.HTTPEmployeesResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.EmployeesData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                },
                "pagination": {
                    "$ref": "#/definitions/entity.Pagination"
                }
            }
        },
        "rest.HTTPErrResp": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.HTTPUserLoginResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.UserLoginData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
...
        "rest.LoginUserData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/rest.UserLoginDataPayload"
                }
            }
        },
...
        "rest.UserLoginData": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/dto.UserLoginDTO"
                }
            }
        },
        "rest.UserLoginDataPayload": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "x-order": "0",
                    "example": "ahmad"
                },
                "password": {
                    "type": "string",
                    "x-order": "1",
                    "example": "sahroni"
                }
            }
        },
        "rest.UserLoginRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.LoginUserData"
                }
            }
        },
...
    }
}